Remove sql.values
The existing implementation suffered from problems with the ordering of
object keys. In most cases, the use of VALUES lists may be replaced with
functions like `jsonb_to_recordset`, so I'm removing `sql.values` for
now. If it proves to be useful in the future, it may come back in some
form.